+++ app/core/GlobalFuntion.php
@@ -40,14 +40,15 @@ if (! function_exists("base_path")) {
   /**
    * Get base path
    *
+   * @param string $insert_path
+   *  Insert string in end of path
    * @return string
    * Base path folder
    */
-  function base_path(): string {
-    return BASEURL;
+  function base_path(string $insert_path = ''): string {
+    return BASEURL . $insert_path;
   }
 }
-
 if (! function_exists("model_path")) {
 
   /**
@@ -58,10 +59,10 @@ if (! function_exists("model_path")) {
    * @return string
    *  Application path folder
    */
-  function model_path(bool $include_basePath = false): string {
+  function model_path(bool $include_basePath = false, string $surfix_path = ''): string {
     return $include_basePath
-      ? (APP_FULLPATH['model'] ?? '')
-      : (APP_PATH['model'] ?? '');
+      ? (APP_FULLPATH['model'] ?? '') . $surfix_path
+      : (APP_PATH['model'] ?? '') . $surfix_path;
   }
 }
 
@@ -75,10 +76,10 @@ if (! function_exists("view_path")) {
    * @return string
    *  Application path folder
    */
-  function view_path(bool $include_basePath = false): string {
+  function view_path(bool $include_basePath = false, string $surfix_path = ''): string {
     return $include_basePath
-      ? (APP_FULLPATH['view'] ?? '')
-      : (APP_PATH['view'] ?? '');
+      ? (APP_FULLPATH['view'] ?? '') . $surfix_path
+      : (APP_PATH['view'] ?? '') . $surfix_path;
   }
 }
 
@@ -92,10 +93,10 @@ if (! function_exists("controllers_path")) {
    * @return string
    *  Application path folder
    */
-  function controllers_path(bool $include_basePath = false): string {
+  function controllers_path(bool $include_basePath = false, string $surfix_path = ''): string {
     return $include_basePath
-      ? (APP_FULLPATH['controllers'] ?? '')
-      : (APP_PATH['controllers'] ?? '');
+      ? (APP_FULLPATH['controllers'] ?? '') . $surfix_path
+      : (APP_PATH['controllers'] ?? '') . $surfix_path;
   }
 }
 
@@ -109,10 +110,10 @@ if (! function_exists("services_path")) {
    * @return string
    *  Application path folder
    */
-  function services_path(bool $include_basePath = false): string {
+  function services_path(bool $include_basePath = false, string $surfix_path = ''): string {
     return $include_basePath
-      ? (APP_FULLPATH['services'] ?? '')
-      : (APP_PATH['services'] ?? '');
+      ? (APP_FULLPATH['services'] ?? '') . $surfix_path
+      : (APP_PATH['services'] ?? '') . $surfix_path;
   }
 }
 
@@ -126,10 +127,10 @@ if (! function_exists("component_path")) {
    * @return string
    *  Application path folder
    */
-  function component_path(bool $include_basePath = false): string {
+  function component_path(bool $include_basePath = false, string $surfix_path = ''): string {
     return $include_basePath
-      ? (APP_FULLPATH['component'] ?? '')
-      : (APP_PATH['component'] ?? '');
+      ? (APP_FULLPATH['component'] ?? '') . $surfix_path
+      : (APP_PATH['component'] ?? '') . $surfix_path;
   }
 }
 
@@ -143,10 +144,10 @@ if (! function_exists("commands_path")) {
    * @return string
    *  Application path folder
    */
-  function commands_path(bool $include_basePath = false): string {
+  function commands_path(bool $include_basePath = false, string $surfix_path = ''): string {
     return $include_basePath
-      ? (APP_FULLPATH['commands'] ?? '')
-      : (APP_PATH['commands'] ?? '');
+      ? (APP_FULLPATH['commands'] ?? '') . $surfix_path
+      : (APP_PATH['commands'] ?? '') . $surfix_path;
   }
 }
 
@@ -160,13 +161,12 @@ if (! function_exists("config_path")) {
    * @return string
    *  Application path folder
    */
-  function config_path(bool $include_basePath = false): string {
+  function config_path(bool $include_basePath = false, string $surfix_path = ''): string {
     return $include_basePath
-      ? (APP_FULLPATH['config'] ?? '')
-      : (APP_PATH['config'] ?? '');
+      ? (APP_FULLPATH['config'] ?? '') . $surfix_path
+      : (APP_PATH['config'] ?? '') . $surfix_path;
   }
 }
-
 
 // string
 
@@ -225,5 +225,21 @@ if (! function_exists("view")) {
 
     // require component
     require_once app_path('view', true) . $view_name . '.template.php';
+  }
+}
+
+// http
+
+if (! function_exists('request')) {
+  function request(): System\Http\Request
+  {
+    return (new System\Http\RequestFactory())->getFromGloball();
+  }
+}
+
+if (! function_exists('response')) {
+  function response($conten = '', int $respone_code = System\Http\Response::HTTP_OK, array $headers = []): System\Http\Response
+  {
+    return (new System\Http\Response($conten, $respone_code, $headers));
   }
 }